+++ src/components/Banner/CardBanner.tsx
@@ -28,6 +28,7 @@ const containerCss = css({
   display: 'flex',
   flexDirection: 'column',
   alignItems: 'center',
+  border: 'none',
 });
 
 const outerContainerCss = css({
@@ -37,8 +38,7 @@ const outerContainerCss = css({
   padding: '0px !', // NOTE: padding 0 필수,
   backgroundOrigin: 'border-box',
   backgroundClip: 'content-box, border-box',
-  backgroundImage:
-    'linear-gradient(token(colors.bg.surface3), token(colors.bg.surface3)), linear-gradient(0deg, #474A5D00 0%, #474A5D 100%)',
+  backgroundImage: 'linear-gradient(#18181D, #18181D), linear-gradient(0deg, #474A5D00 0%, #474A5D 100%)',
 });
 
 const descriptionCss = css({
